Split solution template reload tests into dedicated test classes
