+++ servers/zms/src/test/java/com/yahoo/athenz/zms/SolutionTemplatesReloadStatusTest.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright The Athenz Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.yahoo.athenz.zms;
+
+import org.testng.annotations.Test;
+
+import static org.testng.Assert.assertEquals;
+
+public class SolutionTemplatesReloadStatusTest {
+
+    @Test
+    public void testSolutionTemplatesReloadStatusValues() {
+        SolutionTemplatesReloadStatus[] statuses = SolutionTemplatesReloadStatus.values();
+        assertEquals(statuses.length, 5);
+        assertEquals(statuses[0], SolutionTemplatesReloadStatus.DISABLED);
+        assertEquals(statuses[1], SolutionTemplatesReloadStatus.NOT_MODIFIED);
+        assertEquals(statuses[2], SolutionTemplatesReloadStatus.RELOADED);
+        assertEquals(statuses[3], SolutionTemplatesReloadStatus.RETRY_PENDING);
+        assertEquals(statuses[4], SolutionTemplatesReloadStatus.FAILED);
+    }
+}
